our main plugin now returns a single function that allows options
