+++ src/lib/tailwind/skeleton.cjs
@@ -2,36 +2,18 @@
 // Tailwind Docs: https://tailwindcss.com/docs/plugins
 // Skeleton Docs: https://www.skeleton.dev/guides/tailwind
 
-const plugin = require('tailwindcss/plugin');
+const intellisensePlugin = require('./intellisense.cjs');
+const corePlugin = require('./core.cjs');
 
-// Skeleton Theme Modules
-const themeColors = require('./theme/colors.cjs');
-// Skeleton Design Token Modules
-const tokensBackgrounds = require('./tokens/backgrounds.cjs');
-const tokensBorders = require('./tokens/borders.cjs');
-const tokensBorderRadius = require('./tokens/border-radius.cjs');
-const tokensFills = require('./tokens/fills.cjs');
-const tokensText = require('./tokens/text.cjs');
-const tokensRings = require('./tokens/rings.cjs');
+// The default export is a function that returns an array of plugins
+// and accepts an optional config that determines which plugins are included.
+// By default, all plugins are included.
+module.exports = function (config = { intellisense: true }) {
+	const { intellisense } = config;
+	const plugins = [corePlugin];
 
-module.exports = plugin(
-	({ addUtilities }) => {
-		addUtilities({
-			// Implement Skeleton design token classes
-			...tokensBackgrounds(),
-			...tokensBorders(),
-			...tokensBorderRadius(),
-			...tokensFills(),
-			...tokensText(),
-			...tokensRings()
-		});
-	},
-	{
-		theme: {
-			extend: {
-				// Implement Skeleton theme colors
-				colors: themeColors()
-			}
-		}
-	}
-);
+	// Add the plugin if the option is not explicitly set to false
+	if (intellisense !== false) plugins.push(intellisensePlugin);
+
+	return plugins;
+};
